get_rect_by_dir: Move to `gears.geometry`
Begin to break down `awful.util`
